+++ src/Statistics.php
@@ -13,7 +13,7 @@ use function array_keys;
 use function array_sum;
 use function implode;
 use function number_format;
-use function str_replace;
+use function preg_replace;
 use function strtolower;
 
 /**
@@ -262,7 +262,7 @@ final class Statistics implements JsonSerializable
             return $humans;
         }
 
-        $propertyNormalized = str_replace(' ', '_', strtolower($property));
+        $propertyNormalized = strtolower((string) preg_replace('/[\s_\-]+/', '_', $property));
 
         return $humans[$propertyNormalized] ?? throw new InvalidArgument('Unknown statistics name: "'.$property.'"; expected one of "'.implode('", "', array_keys($humans)).'"');
     }